Move Feature to legacy models file
