+++ api/converters.py
@@ -16,10 +16,10 @@
 import datetime
 from typing import Any
 from google.cloud import ndb  # type: ignore
-from internals import stage_helpers
 
 from internals.core_enums import *
-from internals.core_models import Feature, FeatureEntry, MilestoneSet, Stage
+from internals.core_models import FeatureEntry, MilestoneSet, Stage
+from internals.legacy_models import Feature
 from internals.review_models import Vote, Gate
 from internals import approval_defs
