+++ api/dev_api_test.py
@@ -18,8 +18,9 @@ from google.cloud import ndb  # type: ignore
 
 from api import dev_api
 from internals import stage_helpers
-from internals.core_models import Feature, FeatureEntry, Stage
-from internals.review_models import  Gate
+from internals.core_models import FeatureEntry, Stage
+from internals.legacy_models import Feature
+from internals.review_models import Gate
 
 class DevAPITest(testing_config.CustomTestCase):
