+++ api/features_api_test.py
@@ -20,7 +20,8 @@ import werkzeug.exceptions  # Flask HTTP stuff.
 
 from api import features_api
 from internals import core_enums
-from internals.core_models import Feature, FeatureEntry, MilestoneSet, Stage
+from internals.core_models import FeatureEntry, MilestoneSet, Stage
+from internals.legacy_models import Feature
 from internals import user_models
 from framework import rediscache
