+++ internals/notifier_test.py
@@ -29,7 +29,8 @@ from internals import core_enums
 from internals import notifier
 from internals import stage_helpers
 from internals import user_models
-from internals.core_models import Feature, FeatureEntry, MilestoneSet, Stage
+from internals.core_models import FeatureEntry, MilestoneSet, Stage
+from internals.legacy_models import Feature
 import settings
 
 test_app = flask.Flask(__name__,
