+++ internals/schema_migration.py
@@ -17,7 +17,8 @@ from google.cloud import ndb  # type: ignore
 
 from framework.basehandlers import FlaskHandler
 from internals import approval_defs, stage_helpers
-from internals.core_models import Feature, FeatureEntry, MilestoneSet, Stage
+from internals.core_models import FeatureEntry, MilestoneSet, Stage
+from internals.legacy_models import Feature
 from internals.review_models import Activity, Approval, Comment, Gate, Vote
 from internals.core_enums import *
