+++ pages/guide_test.py
@@ -21,7 +21,8 @@ from google.cloud import ndb  # type: ignore
 from framework import rediscache
 from internals import core_enums
 from internals import stage_helpers
-from internals.core_models import Feature, FeatureEntry, MilestoneSet, Stage
+from internals.core_models import FeatureEntry, MilestoneSet, Stage
+from internals.legacy_models import Feature
 from internals.review_models import Gate
 from pages import guide
